extract color and text commands into their own files in functions sub-directory
